requirements.txt
gitpython
kopf
pandas
paramiko
jupyter
notebook